+++ CHANGES.md
@@ -1,7 +1,39 @@
-# NGI Python KOF Parser
+# NGI Python KOF Parser Package
 
-## Version 0.0.1
+## Version 0.0.6
+_2022-02-21_
 
--- Initial version
+Add
+
+- Support for more methods. Complete list:        
+   - RO: 2251
+   - RW: 2401
+   - SA: 2402
+   - TP: 2403
+   - SS: 2405
+   - RP: 2406
+   - CPT: 2407
+   - RS: 2409
+   - SR: 2410
+   - SPT: 2411
+   - RCD: 2412
+   - PZ: 2413
+   - PT: 2414 
+   - SVT: 2415
+   - INC: 2417
+   - TOT: 2418
+- Support parsing of letter tema codes F, VB and GVR.
+- KOF Writer for creating KOF files.
+- Add coordinate system handling in the admin block (01).
+
+
+## Version 0.0.5
+_2022-02-10_
+
+
+## Version 0.0.3
+_2022-01-31_
+
+- Initial version